+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added INGXLogger.partialUpdateConfig that updates only some fields of the config
 - Added partialConfig parameter to CustomNGXLoggerService.getNewInstance
+- Added context to INGXLoggerConfig, context is written in {} to the log for every log messages
 
 ### Changed
 - INGXLoggerConfig.level is now optional (ERROR is the default value)